+++ src/app/sitemap.ts
@@ -33,6 +33,12 @@ export default function sitemap(): MetadataRoute.Sitemap {
       priority: 0.8,
     },
     {
+      url: 'https://ourtrips.to/pricing',
+      lastModified: new Date(),
+      changeFrequency: 'weekly',
+      priority: 0.8,
+    },
+    {
       url: 'https://ourtrips.to/changelog',
       lastModified: new Date(),
       changeFrequency: 'weekly',